scripts/requirements.txt
duckdb
numpy
pandas
python-dotenv
spotipy